move validated cluster to the controller which will be the central dependency for subsequent steps
